The prescibred Field functionality of the code was changed. It does not apply the time dependency of the prescribed field anymore, but uses the field at t=0 for the whole simulation. The Python function defining the prescribed field is not evaluated at each timestep anymore, which makes the code much faster.
I added generatePrescribedFields and rememberPrescribedFields methods to VectorPatch, which in turn call the corresponding
methodes in Electromagn. In the Field class, i added a addFrom function, which is similar to copyFrom. It adds all values of
a target field to the object.
The generatePrescribedFields calls the Python profile via the original applyPrescribedField method and saves
the result to a new member of the extTimeFields struct.
The remember method replaces the applyPrescribedField method in Smilei.cpp. Like applyPrescribedField, it first restores
the original fields values and then adds the saved values of the prescibed Field via the new addFrom method.
In Electromagn, I added a member prescibedField, to which the constant prescribed field is stored. It is initialized alongside
savedField in ElectroMagnFactory.
	new file:   make_verbose_output.txt
	modified:   src/ElectroMagn/ElectroMagn.cpp
	modified:   src/ElectroMagn/ElectroMagn.h
	modified:   src/ElectroMagn/ElectroMagnFactory.h
	modified:   src/Field/Field.h
	modified:   src/Patch/VectorPatch.cpp
	modified:   src/Patch/VectorPatch.h
	modified:   src/Smilei.cpp